Move dependencies merging to a separate module
