+++ eo-phi-normalizer/app/Main.hs
@@ -26,7 +26,7 @@
 
 module Main (main) where
 
-import Control.Monad (foldM, forM, unless, when)
+import Control.Monad (forM, unless, when)
 import Data.Foldable (forM_)
 
 import Control.Exception (Exception (..), SomeException, catch, throw)
@@ -37,8 +37,9 @@ import Data.Text.Internal.Builder (toLazyText)
 import Data.Text.Lazy as TL (unpack)
 import Data.Yaml (decodeFileThrow)
 import GHC.Generics (Generic)
-import Language.EO.Phi (AlphaIndex (AlphaIndex), Attribute (Alpha), Binding (..), Bytes (Bytes), Function (..), Object (Formation), Program (Program), parseProgram, printTree)
+import Language.EO.Phi (Bytes (Bytes), Object (Formation), Program (Program), parseProgram, printTree)
 import Language.EO.Phi.Dataize
+import Language.EO.Phi.Dependencies
 import Language.EO.Phi.Metrics.Collect as Metrics (getProgramMetrics)
 import Language.EO.Phi.Metrics.Data as Metrics (ProgramMetrics (..), splitPath)
 import Language.EO.Phi.Report.Data (Report'InputConfig (..), Report'OutputConfig (..), ReportConfig (..), ReportItem (..), makeProgramReport, makeReport)
@@ -346,79 +347,6 @@ getMetrics :: Maybe String -> Maybe FilePath -> IO ProgramMetrics
 getMetrics bindingsPath inputFile = do
   program <- getProgram inputFile
   either throw pure (getMetrics' program bindingsPath)
-
--- ** Merging programs with dependencies
-
-bindingAttr :: Binding -> Maybe Attribute
-bindingAttr (AlphaBinding a _) = Just a
-bindingAttr (EmptyBinding a) = Just a
-bindingAttr (DeltaBinding _) = Just (Alpha (AlphaIndex "Δ"))
-bindingAttr DeltaEmptyBinding = Just (Alpha (AlphaIndex "Δ"))
-bindingAttr LambdaBinding{} = Just (Alpha (AlphaIndex "λ"))
-bindingAttr MetaBindings{} = Nothing
-bindingAttr MetaDeltaBinding{} = Nothing
-
-zipBindings :: [Binding] -> [Binding] -> ([Binding], [(Binding, Binding)])
-zipBindings xs ys = (xs' <> ys', collisions)
- where
-  as = map bindingAttr xs
-  bs = map bindingAttr ys
-
-  xs' = [x | x <- xs, bindingAttr x `notElem` bs]
-  ys' = [y | y <- ys, bindingAttr y `notElem` as]
-  collisions =
-    [ (x, y)
-    | x <- xs
-    , y <- ys
-    , bindingAttr x == bindingAttr y
-    ]
-
-isPackage :: [Binding] -> Bool
-isPackage = any isPackageBinding
-
-isPackageBinding :: Binding -> Bool
-isPackageBinding (LambdaBinding (Function "Package")) = True
-isPackageBinding _ = False
-
-mergeBinding :: Binding -> Binding -> Either String Binding
-mergeBinding (AlphaBinding a (Formation xs)) (AlphaBinding b (Formation ys))
-  | a == b = AlphaBinding a . Formation <$> mergeBindings xs ys
-mergeBinding x y | x == y = return x
-mergeBinding x y =
-  Left $
-    concat @[]
-      [ "conflict when adding dependencies (trying to merge non-formations)"
-      , printTree x
-      , printTree y
-      ]
-
-mergeBindings :: [Binding] -> [Binding] -> Either String [Binding]
-mergeBindings xs ys
-  | isPackage xs && isPackage ys = do
-      case zipBindings xs ys of
-        (zs, collisions) -> do
-          ws <- mapM (uncurry mergeBinding) collisions
-          return (zs <> ws)
-  | otherwise =
-      Left $
-        concat @[]
-          [ "conflict when adding dependencies (trying to merge non-Package formations "
-          , printTree (Formation xs)
-          , printTree (Formation ys)
-          , " )"
-          ]
-
-deepMerge :: Program -> Program -> Either String Program
-deepMerge (Program xs) (Program ys) = Program <$> mergeBindings (mkPackage xs) (mkPackage ys)
- where
-  mkPackage bs
-    | isPackage bs = bs
-    -- FIXME: check if lambda attribute exists and throw error!
-    | otherwise = LambdaBinding (Function "Package") : bs
-
-deepMergePrograms :: [Program] -> Either String Program
-deepMergePrograms [] = Right (Program [])
-deepMergePrograms (p : ps) = foldM deepMerge p ps
 
 -- * Main
